# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TechPulse AI API",
    "version": "1.7.0",
    "description": "Canonical HTTP contract for the Vietnamese-first TechPulse AI MVP. Browser API access is same-origin only; external content is untrusted data; no response contains source full text or media binary."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin API. Production sends no CORS allow-origin/credentials headers; any future exception requires an exact reviewed scheme+host allowlist."
    }
  ],
...
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/admin/ingestion-jobs": {
      "get": {
        "tags": [
          "Admin Jobs"
        ],
        "operationId": "listIngestionJobs",
        "x-persistence": "mongo",
        "summary": "List durable ingestion jobs",
        "parameters": [
          {
            "$ref": "#/components/parameters/CursorQuery"
          },
          {
            "$ref": "#/components/parameters/LimitQuery"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/JobStatus"
            }
          },
          {
            "name": "sourceId",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionJobListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "post": {
        "tags": [
          "Admin Jobs"
        ],
        "operationId": "createIngestionJob",
        "x-persistence": "mongo",
        "summary": "Trigger bounded ingestion using the shared job service",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfTokenHeader"
          },
          {
            "$ref": "#/components/parameters/BrowserOriginHeader"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKeyHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestionJobCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Job accepted; inspect returned state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionJobResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/v1/admin/ingestion-jobs/{jobId}": {
      "get": {
        "tags": [
          "Admin Jobs"
        ],
        "operationId": "getIngestionJob",
        "x-persistence": "mongo",
        "summary": "Get one ingestion job",
        "parameters": [
          {
            "$ref": "#/components/parameters/JobIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionJobResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/v1/admin/ingestion-jobs/{jobId}/retries": {
      "post": {
        "tags": [
          "Admin Jobs"
        ],
        "operationId": "retryIngestionJob",
        "x-persistence": "mongo",
        "summary": "Create a bounded retry attempt",
        "parameters": [
          {
            "$ref": "#/components/parameters/JobIdPath"
          },
          {
            "$ref": "#/components/parameters/CsrfTokenHeader"
          },
          {
            "$ref": "#/components/parameters/BrowserOriginHeader"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKeyHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRetryRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Retry accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionJobResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/v1/admin/ingestion-jobs/{jobId}/cancellation": {
      "post": {
        "tags": [
          "Admin Jobs"
        ],
        "operationId": "cancelIngestionJob",
        "x-persistence": "mongo",
        "summary": "Cancel a queued job or request safe stop",
        "parameters": [
          {
            "$ref": "#/components/parameters/JobIdPath"
          },
          {
            "$ref": "#/components/parameters/CsrfTokenHeader"
          },
          {
            "$ref": "#/components/parameters/BrowserOriginHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCancellationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionJobResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-techpulse_session",
        "description": "Opaque server-side session token. The cookie is host-only and always emitted with Secure, HttpOnly, Path=/ and SameSite=Lax; Domain is forbidden. Every cookie-authenticated mutation also passes the exact same-origin browser Origin gate and CSRF validation."
      },
      "cronBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Vercel CRON_SECRET; never accepted from a browser session."
      }
    },
...
    "parameters": {
      "BrowserOriginHeader": {
        "name": "Origin",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uri",
          "maxLength": 2048
        },
        "description": "Browser origin. Production requires an exact normalized scheme+host+effective-port match with the configured public application origin; suffix, wildcard and reflected matching are forbidden."
      },
      "CsrfTokenHeader": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 32,
          "maxLength": 256
        },
        "description": "Session-bound CSRF token. The common browser ingress also requires an exact same-origin Origin value on every cookie-authenticated mutation."
      },
      "IdempotencyKeyHeader": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{7,127}$"
        },
        "description": "Scoped by authenticated actor/session and canonical request hash. Reusing a key for a different intent returns 409 idempotency_mismatch. The guarantee is 24 hours for grounded answers and at least 14 days for manual job/governance mutations; scheduled identities remain deterministic by target and period."
      },
      "CursorQuery": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 1000
        },
        "description": "Opaque cursor; clients must not parse it."
      },
      "LimitQuery": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
...
      "JobIdPath": {
        "name": "jobId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "bad_request",
                "message": "Malformed request",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "unauthorized",
                "message": "Authentication required",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permission or invalid CSRF token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "forbidden",
                "message": "Insufficient permission or invalid CSRF token",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "not_found",
                "message": "Resource not found",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      },
      "Conflict": {
        "description": "State, duplicate, lease, or idempotency conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "conflict",
                "message": "State, duplicate, lease, or idempotency conflict",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "Semantic validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "validation_error",
                "message": "Semantic validation failed",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limit exceeded",
        "headers": {
          "Retry-After": {
            "description": "Seconds until retry is permitted",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "rate_limit_exceeded",
                "message": "Rate limit exceeded",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "Request target or JSON payload exceeds the global ingress limit",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "payload_too_large",
                "message": "Request exceeds the allowed size",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      },
      "UnsupportedMediaType": {
        "description": "JSON route requires application/json with identity content encoding",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "unsupported_media_type",
                "message": "Unsupported request content type or encoding",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected internal failure",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "internal_error",
                "message": "Unexpected internal failure",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "Required dependency is temporarily unavailable",
        "headers": {
          "Retry-After": {
            "description": "Seconds until a retry may succeed",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "service_unavailable",
                "message": "Required dependency is temporarily unavailable",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "ConnectorType": {
        "type": "string",
        "enum": [
          "rss",
          "arxiv",
          "hacker-news"
        ]
      },
...
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "partial",
          "failed",
          "cancelled"
        ]
      },
...
      "CursorMeta": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "hasNext",
          "nextCursor"
        ],
        "properties": {
          "hasNext": {
            "type": "boolean"
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "SafeError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "retryable",
          "occurredAt"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "upstreamStatus": {
            "type": [
              "integer",
              "null"
            ]
          }
        }
      },
      "JobCounters": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "fetched",
          "created",
          "updated",
          "duplicate",
          "skipped",
          "failed"
        ],
        "properties": {
          "fetched": {
            "type": "integer",
            "minimum": 0
          },
          "created": {
            "type": "integer",
            "minimum": 0
          },
          "updated": {
            "type": "integer",
            "minimum": 0
          },
          "duplicate": {
            "type": "integer",
            "minimum": 0
          },
          "skipped": {
            "type": "integer",
            "minimum": 0
          },
          "failed": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "IngestionJob": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "idempotencyKey",
          "sourceId",
          "connectorType",
          "expectedSourcePolicyVersion",
          "trigger",
          "status",
          "attempt",
          "availableAt",
          "leaseGeneration",
          "batchSize",
          "counters",
          "error",
          "createdAt",
          "startedAt",
          "finishedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "idempotencyKey": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "connectorType": {
            "$ref": "#/components/schemas/ConnectorType"
          },
          "expectedSourcePolicyVersion": {
            "type": "integer",
            "minimum": 1,
            "description": "Source policy/config version captured before external fetch and required by the final article/checkpoint transaction."
          },
          "trigger": {
            "type": "string",
            "enum": [
              "cron",
              "admin",
              "retry"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "attempt": {
            "type": "integer",
            "minimum": 1
          },
          "availableAt": {
            "type": "string",
            "format": "date-time"
          },
          "leaseGeneration": {
            "type": "integer",
            "minimum": 0
          },
          "batchSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "parentJobId": {
            "type": [
              "string",
              "null"
            ]
          },
          "counters": {
            "$ref": "#/components/schemas/JobCounters"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SafeError"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "finishedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "IngestionJobCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "sourceId"
        ],
        "properties": {
          "sourceId": {
            "type": "string"
          },
          "batchSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          }
        }
      },
...
      "JobRetryRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["reasonCode"],
        "properties": {
          "reasonCode": { "const": "job_retry_requested" }
        }
      },
      "JobCancellationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["reasonCode"],
        "properties": {
          "reasonCode": { "const": "job_cancel_requested" }
        }
      },
...
      "IngestionJobResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/IngestionJob"
          }
        }
      },
      "IngestionJobListResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IngestionJob"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/CursorMeta"
          }
        }
      },
...
      "ErrorDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "message",
          "code"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "bad_request",
          "validation_error",
          "unauthorized",
          "csrf_invalid",
          "forbidden",
          "not_found",
          "conflict",
          "idempotency_mismatch",
          "invalid_state_transition",
          "source_policy_blocked",
          "rate_limit_exceeded",
          "payload_too_large",
          "unsupported_media_type",
          "provider_unavailable",
          "database_unavailable",
          "service_unavailable",
          "internal_error"
        ],
        "description": "Stable machine-readable taxonomy. Operations may document a narrower subset; clients must not branch on message text."
      },
      "ErrorObject": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          },
          "requestId": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorObject"
          }
        }
      }
    }
  }
}